WormlikeChain

WormlikeChain.BeadDefinition

WormlikeChain.Chain

WormlikeChain.ChainSystem

WormlikeChain.SpecificForce

WormlikeChain.SpecificForce

WormlikeChain.next_beadid

WormlikeChain.prev_beadid

WormlikeChain.randbits_2x64

WormlikeChain.randbits_4x32

WormlikeChain.randn_2x64

WormlikeChain.randn_32

WormlikeChain.randn_4x32

WormlikeChain.BeadDefinition — Type

The potential and number of dimentions of a bead

WormlikeChain.Chain — Method

zero velocities by default

WormlikeChain.ChainSystem — Type

A collection of Chains and SpecificForces

WormlikeChain.SpecificForce — Type

A force that acts on a specific list of beads

WormlikeChain.SpecificForce — Method

Create a specific force pe

WormlikeChain.next_beadid — Method

Return the bead id of the next bead on a chain

chainbounds is a tuple of at least 2 Ints, chainbounds[1] is 1 chainbounds[end] is the total number of beads + 1 the other element in chainbounds are the bead ids of the start of a new chain

WormlikeChain.prev_beadid — Method

Return the bead id of the previous bead on a chain

chainbounds is a tuple of at least 2 Ints, chainbounds[1] is 1 chainbounds[end] is the total number of beads + 1 the other element in chainbounds are the bead ids of the start of a new chain

WormlikeChain.randbits_2x64 — Method

Return a tuple of 2 random UInt64 Uses the philox random CBRNG 4x32_10

WormlikeChain.randbits_4x32 — Method

Return a tuple of 4 random UInt32 Uses the philox random CBRNG 4x32_10

WormlikeChain.randn_2x64 — Method

Return a tuple of 2 normal random Float64 Uses the philox random CBRNG 4x32_10 and box-muller transform

WormlikeChain.randn_32 — Method

Return a tuple of n normal random Float32, incrementing ctr2 for each 4. Uses the philox random CBRNG 4x32_10 and box-muller transform

WormlikeChain.randn_4x32 — Method

Return a tuple of 4 normal random Float32 Uses the philox random CBRNG 4x32_10 and box-muller transform